+++ tests/LinkDotNet.StringBuilder.UnitTests/ValueStringBuilder.Trim.Tests.cs
@@ -46,4 +46,37 @@ public class ValueStringBuilderTrimTests
 
         valueStringBuilder.ToString().Should().Be(expected);
     }
+
+    [Fact]
+    public void GivenString_WhenTrimStartCharacter_ThenShouldRemoveCharacter()
+    {
+        using var valueStringBuilder = new ValueStringBuilder();
+        valueStringBuilder.Append("HHeeHH");
+
+        valueStringBuilder.TrimStart('H');
+
+        valueStringBuilder.ToString().Should().Be("eeHH");
+    }
+
+    [Fact]
+    public void GivenString_WhenTrimEndCharacter_ThenShouldRemoveCharacter()
+    {
+        using var valueStringBuilder = new ValueStringBuilder();
+        valueStringBuilder.Append("HHeeHH");
+
+        valueStringBuilder.TrimEnd('H');
+
+        valueStringBuilder.ToString().Should().Be("HHee");
+    }
+
+    [Fact]
+    public void GivenString_WhenTrimCharacter_ThenShouldRemoveCharacter()
+    {
+        using var valueStringBuilder = new ValueStringBuilder();
+        valueStringBuilder.Append("HHeeHH");
+
+        valueStringBuilder.Trim('H');
+
+        valueStringBuilder.ToString().Should().Be("ee");
+    }
 }